fix(grab): sync menu by PUSHing it, not notify (UnsupportedMenuSync)
The "Sync menu" buttons called notifyMenuUpdate (POST .../menu/notification),
which Grab rejects for our self-serve-activated stores:
  400 {"target":"UnsupportedMenuSync","reason":"invalid_argument"}
These stores aren't on the notify→pull model — they're push: the partner PUTs
the menu to Grab (same family as the live 86/availability push).

- New lib/grab-menu-outlet.buildOutletGrabMenu(): the single per-outlet menu
  builder (channel-visible products + per-outlet service hours + 86 list),
  shared by both menu directions so a push is byte-for-byte what a pull serves.
- /api/pos/grab/sync now builds that menu and PUTs it via updateMenu
  (PUT /partner/v1/menu) instead of notifyMenuUpdate. Surfaces Grab's exact
  error on failure. Service-role read for per-outlet data; still staff-gated.
- get-menu webhook refactored onto the shared builder (no behaviour change;
  preserves the pull path Tamarind's self-serve activation will use).
- Page copy: "pushed to GrabFood" instead of "Grab will re-pull".
